+++ Realm/RLMSyncUtil.mm
@@ -18,7 +18,6 @@
 
 #import "RLMSyncUtil_Private.hpp"
 
-#import "RLMJSONModels.h"
 #import "RLMObject_Private.hpp"
 #import "RLMRealmConfiguration+Sync.h"
 #import "RLMRealmConfiguration_Private.hpp"
@@ -31,8 +30,6 @@
 
 #import "sync/sync_user.hpp"
 
-RLMIdentityProvider const RLMIdentityProviderAccessToken = @"_access_token";
-
 NSString *const RLMSyncErrorDomain = @"io.realm.sync";
 NSString *const RLMSyncAuthErrorDomain = @"io.realm.sync.auth";
 NSString *const RLMAppErrorDomain = @"io.realm.app";
@@ -40,21 +37,8 @@ NSString *const RLMAppErrorDomain = @"io.realm.app";
 NSString *const kRLMSyncPathOfRealmBackupCopyKey            = @"recovered_realm_location_path";
 NSString *const kRLMSyncErrorActionTokenKey                 = @"error_action_token";
 
-NSString *const kRLMSyncAppIDKey                = @"app_id";
-NSString *const kRLMSyncDataKey                 = @"data";
-NSString *const kRLMSyncErrorJSONKey            = @"json";
 NSString *const kRLMSyncErrorStatusCodeKey      = @"statusCode";
-NSString *const kRLMSyncIdentityKey             = @"identity";
-NSString *const kRLMSyncIsAdminKey              = @"is_admin";
-NSString *const kRLMSyncNewPasswordKey          = @"new_password";
-NSString *const kRLMSyncPasswordKey             = @"password";
-NSString *const kRLMSyncPathKey                 = @"path";
-NSString *const kRLMSyncProviderKey             = @"provider";
-NSString *const kRLMSyncProviderIDKey           = @"provider_id";
-NSString *const kRLMSyncRegisterKey             = @"register";
-NSString *const kRLMSyncTokenKey                = @"token";
 NSString *const kRLMSyncUnderlyingErrorKey      = @"underlying_error";
-NSString *const kRLMSyncUserIDKey               = @"user_id";
 
 #pragma mark - C++ APIs
 
@@ -90,35 +74,6 @@ std::shared_ptr<SyncSession> sync_session_for_realm(RLMRealm *realm) {
 CocoaSyncUserContext& context_for(const std::shared_ptr<realm::SyncUser>& user)
 {
     return *std::static_pointer_cast<CocoaSyncUserContext>(user->binding_context());
-}
-
-NSError *make_auth_error_bad_response(NSDictionary *json) {
-    return [NSError errorWithDomain:RLMSyncAuthErrorDomain
-                               code:RLMSyncAuthErrorBadResponse
-                           userInfo:json ? @{kRLMSyncErrorJSONKey: json} : nil];
-}
-
-NSError *make_auth_error_http_status(NSInteger status) {
-    return [NSError errorWithDomain:RLMSyncAuthErrorDomain
-                               code:RLMSyncAuthErrorHTTPStatusCodeError
-                           userInfo:@{kRLMSyncErrorStatusCodeKey: @(status)}];
-}
-
-NSError *make_auth_error_client_issue() {
-    return [NSError errorWithDomain:RLMSyncAuthErrorDomain
-                               code:RLMSyncAuthErrorClientSessionError
-                           userInfo:nil];
-}
-
-NSError *make_auth_error(RLMSyncErrorResponseModel *model) {
-    NSMutableDictionary<NSString *, NSString *> *userInfo = [NSMutableDictionary dictionaryWithCapacity:2];
-    if (NSString *description = model.title) {
-        [userInfo setObject:description forKey:NSLocalizedDescriptionKey];
-    }
-    if (NSString *hint = model.hint) {
-        [userInfo setObject:hint forKey:NSLocalizedRecoverySuggestionErrorKey];
-    }
-    return [NSError errorWithDomain:RLMSyncAuthErrorDomain code:model.code userInfo:userInfo];
 }
 
 NSError *make_sync_error(RLMSyncSystemErrorKind kind, NSString *description, NSInteger code, NSDictionary *custom) {