+++ RealmSwift/Sync.swift
@@ -27,21 +27,6 @@ import Realm.Private
 public typealias User = RLMUser
 
 /**
- An immutable data object representing information retrieved from MongoDB
- Realm about a particular user.
-
- - see: `RLMUserInfo`
- */
-public typealias UserInfo = RLMUserInfo
-
-/**
- An immutable data object representing an account belonging to a particular user.
-
- - see: `UserInfo`, `RLMUserAccountInfo`
- */
-public typealias UserAccountInfo = RLMUserAccountInfo
-
-/**
  A singleton which configures and manages MongoDB Realm synchronization-related
  functionality.
 
@@ -327,7 +312,7 @@ public extension User {
     /**
      Create a sync configuration instance.
 
-     - parameter partitionValue: FIXME
+     - parameter partitionValue: The `BSON` value the Realm is partitioned on.
      - parameter cancelAsyncOpenOnNonFatalErrors: By default, Realm.asyncOpen()
      swallows non-fatal connection errors such as a connection attempt timing
      out and simply retries until it succeeds. If this is set to `true`, instead